// app/dashboard/models/CorrespondenenceDetails.model.ts
export class CorrespondenenceDetailsModel {
  CorrFolderName: string;
  AttachID: string;
  AttachCorrID: string;
  AttachCorrCoverID: string;
  AttachCorrAttachmentsID: string;
  VolumeID: string;
  SubWorkID: string;
  AttachCorrMiscID: string;
  countconnections: string;
  SubWork_Status: string;
  ID: string;
  UserID: string;
  TransferUserID: string;
  Status: string;
  CCstatus: string;
  holdSecretaryID: string;
  TransferType: string;
  isCC: string;
  perform07: string;
  WasOpened: string;
  AttachementsDocCount: string;

}